chore(CODEOWNERS): update code ownership
This repository has been more or less unmaintained for a little while,
even though it is very much still in use by Einride. One reason for
that is that the team who was assigned ownership has a lot of
ownership already and was coping to keep up.

Jassob and Max Ekman have expressed interest in continuing to maintain
the repository and interact with the community.