feat(cantool): generate raw setters and getters for physical signals
For physical signals that have value table definitions we generate
getters and setters that operate on the raw signal value, to give
users access to check if signal values correspond to a named constant,
or to write named constants to the signal.

However, there are use-cases (especially for sending test data) for
writing raw values to any physical signal and therefore this commit
adds generation of raw getters and setters to all physical signals.