fix: bump Golang version to go1.20
One of our dependencies requires Go 1.20, which means that the project
is not compilable with Go 1.19. This commit updates the module files
to demand the correct version.

This is not a breaking change since the repository was not buildable
with Go 1.19 and thus we should not be breaking anything for anybody.